Add option to override number of antennas
Currently the ObservationInfo structure gets the number of antennas
from the global object VCS_OBSERVATION_INFO. And it is always 128.

With the `-R` option we can override that number through the commandline.
This is necessary to image MWAX voltages.

In the future we can get this information from the metadata, when we move
metadata reading in AstroIO. But for now, this will do.