Finished the first implementation of images RFI flagging.
The RMS for each image is estimated with a 2-stage process implemented
in the compute_images_rms kernel. Basically, outlier pixels are removed
from the image RMS computation. Then, RMS is computed on the remaining
pixels.

The vector of RMS is then transferred to CPU to compute which images
are going to be flagged. Given the small data and computation, it is
more convenient to do it on the CPU.

As a mitigation, images that are flagged can be cleared, that is,
their pixels set to zero so that they do not cause false detection.
This can be done with the `clear_flagged_images_gpu` function.