Fixes long-standing "dedispersion bug".
The bug was not actually in the dedispersion code. Instead, I assumed
a fixed mapping between loop index values and thread ids when
distributing input voltage objects across multiple OpenMP threads
in the parallelised for-loop in the `run` method.

I wrongly thought OpenMP would assign index 0 to thread 0, index 1
to thread 1, and so on. That was the case on NVIDIA systems
(gcc compiler), but not on HIP (clang).

Furthermore, I was copying the mapping and solutions array from
a GPU memory to other GPUs memory, and not sure this always works.
Hence, I now create copies on CPU which then are moved to GPU memory.