Implements support for adding images to the dynamic spectrum using GPU.
However, this implementation is not complete as the `MemoryBuffer` class does not yet handle Managed Memory allocation.

Because there a dynamic spectrum receives contributions from all coarse channels, that are distributed across multiple GPUs,
it cannot be allocated on a single GPU. All GPUs must be able to write to this same memory location (the dynamic spectrum),
to add their contributions. Hence, it is better to allocate the dynamic spectrum in Managed Memory. It will make it also
easier to process it for RFI flagging as it is directly accessible to CPU.