Implements timestep flagging.
This uses the same principle Marcin defined: "total power" or, better
rms of power across the band within one time step. However, in the
pipeline we only process one coarse channel at a time, so it will
have to do.

We compute the rms for each time step in a second, and if the rms
is greater than 3 times the minimum rms, then we flag the timestep.
The above may seem very aggressive, but I find that rms does not
vary very much across "good" timesteps. On average, rms of rms is
1.x on a good second of data. In bad seconds, rms of rms goes up to
16 and above. Hence, the usual threshold based on median and rms
of rms is not really effective in eliminating all RFI.

Furthermore, if more than 4 timesteps are flagged, we also flag
all the timesteps in between them.