Implements a better RFI flagging, mostly on CPU.
The flag is the same, -f. However, I have tested a lower tollerance threshold, 5, and works well on several test observations. So, my default will be -f 5.

The RFI mitigation strategy computes the RMS and MEAN values for the central region (100x100) of each image. Then, it uses these values to flag anomalous images within a 1s / coarse channel block in three ways:

Discarding images whose RMS OR MEAN values are outside the tollerance limit established across all images in the block.
Compute the MEADIAN RMS and MEDIAN of MEAN for all timesteps in a block, and discard timesteps where either one of them is outside tollerance limits. This is good for broadband RFI.
Compute the MEADIAN RMS and MEDIAN of MEAN for all channels in a block, and discard channels where either one of them is outside tollerance limits. This is good for narrowband RFI.

Tollerance limit for RMS values associated with timesteps and channels are stricter within a block . First, RMS can only be a positive number. Hence, the condition for flagging is hardcoded to: value > 3 * min_rms, where min_rms is the minimum RMS value across all timesteps (or channels). This is because with broadband RFI, many timesteps in the block have bad RMS, and median representative of good values. The minimum value I find it more useful.

Then, to make the system more robust, one could extend those checks using a separate history for timesteps, channels, and single images RMS and Median values.